removed qubit, trimmed target
