+++ src/sympleq/core/circuits/gates.py
@@ -15,7 +15,7 @@ from sympleq.core.circuits.random_symplectic import (
 )
 from sympleq.core.circuits.find_symplectic import map_paulisum_to_target_tableau
 from sympleq.core.paulis.constants import DEFAULT_QUDIT_DIMENSION
-from sympleq.core.circuits.target import get_phase_vector
+from sympleq.core.circuits.helpers_solve_from_target import get_phase_vector
 from sympleq.core.paulis.pauli_string import PauliString
 from sympleq.core.paulis.pauli_sum import PauliSum
